Modify HTML and JS
Modified index.html to add onclick functions to the gameboard grid's divs
Modified script.js to overhaul the game mechanics. Got rid of the playRound function and moved most of that code into updateBoard function. I had to do this as nothing
was working as intended after I added the frontend to the web page. All of my javascript is probably awful but it works, for now.